Move lambda code into src folder.
