+++ lambda/src/__mocks__/yamljs.js
