+++ lambda/src/hearts/__tests__/testGetHearts.js
