+++ lambda/src/hearts/hearts.js
