+++ lambda/src/migrate_auth/__tests__/migrate_auth.test.js
