+++ lambda/src/migrate_auth/cognito.js
