+++ lambda/src/migrate_auth/migrate_auth.js
