+++ lambda/src/utils/__mocks__/database.js
